requirements.txt
flask
anthropic
python-dotenv
pyodbc
pytest
pandas
numpy
faker
chainlit
langchain-anthropic
langchain-core == 0.2.19
pydantic==2.10.1
ipykernel
aiohttp